data loader telegram fixed

+++ torch_geometric_signed_directed/data/directed/Telegram.py
@@ -58,7 +58,7 @@ class Telegram(InMemoryDataset):
         indices = torch.from_numpy(indices).long()
         features = torch.from_numpy(rs.normal(0, 1.0, (s_A.shape[0], 1))).float()
 
-        data = Data(x=features, edge_index=indices, edge_weight=values, y=label)
+        data = Data(x=features, edge_index=indices, edge_weight=torch.FloatTensor(values), y=label)
         data = node_class_split(data, train_size_per_class=train_ratio, val_size_per_class=val_ratio)
         if self.pre_transform is not None:
             data = self.pre_transform(data)
